Backend/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
pymysql==1.1.0
cryptography==41.0.0